Changed the way questions are generated. Conventionally this is done by taking the shortest prefix of R and concatenating the set of suffixes from B. If B is a leaf node however, the question is simply the prefix of R, which is already accepted, so the merge is carried out regardless. Now, if this is the case, a question is generated that takes the shortest prefix up to B (not R), and concatenates the set of labels that are adjacent to R.
Also, if there exists an edge R->R , it is not taken into account by conventional question generation. It does not 
belong to the shortest path from q0 to R, and does not belong to the set of suffixes. This has been rectified, so that
if there is such an edge (potentially with several labels), the question is formulated as:
prefix X setOfLabels(R->R) X Suffixes(B)

Finally, if R and B are neighbours, the edge R -a-> B will become R -a-> R. I.e. a loop will be created. Conventionally the 
questions to not check whether a new loop is acceptable (i.e. whether the sequence R-a->R-a->R is acceptable. This has been added.

git-svn-id: svn://svn.code.sf.net/p/statechum/code/XMachineTool/trunk@35 4b8d0e13-f52a-0410-bc0e-8c3b9955e25a