Call graph gets constructed.
Need to record method signatures in traces as well if we are going to use call graph to answer questions. The reason for this is that a class can have several overriding methods with the same name, which can only be distinguished by their signatures. This has been added to the implementation.

git-svn-id: svn://svn.code.sf.net/p/statechum/code/XMachineTool/trunk@151 4b8d0e13-f52a-0410-bc0e-8c3b9955e25a