Seems like finished improving the speed of learning.
Replaced ArrayMapWithSearchAndCounter with a facade
MapWithSearchAndCounter that can be applied anywhere. It is now possible
to dynamically select the type of transition matrix depending on the
number of elements, which helps a lot. Tests of learning from a large
PTA now run in a few min instead of an hour.